seeder/requirements.txt
psycopg2-binary==2.9.10
cassandra-driver==3.29.2
faker==34